fixed POST and GET
The database for the backend was not properly displaying the reviews so I had to fix it. But now it should be fine. I now need to implement the front end for this. I know no one is really going to see my project in full detail and I could technically leave this as is, but Id like to think whoever is reading this might give it a chance :)