Final git fixes and minor front end fixes
had problems with the front end being reported as a submodule which affected backups. After a tedious amount of time I was able to fix it. Now I have to fix the modify, add a page where the user can add a new entry properly. I also made it so that the scores would not constantly disappear randomly.